scenario1_single_server/baseline_agent.py
from strands import Agent
task = "write Python code to differentiate an equation"
model = "anthropic.claude-3-sonnet-20240229-v1:0"
agent = Agent(model=model,system_prompt="you are a helpful coding assistant")
response = agent(task)
print(response)

#Output will be like this:
"""
To differentiate an equation in Python, we need to use symbolic computation libraries like SymPy. SymPy is a Python library for symbolic mathematics, and it can be used to perform various mathematical operations, including differentiation.

Here's an example code that demonstrates how to differentiate an equation using SymPy:

```python
from sympy import symbols, diff

# Define the symbolic variables
x = symbols('x')

# Define the equation
equation = x**3 + 2*x**2 - 3*x + 5

# Differentiate the equation
derivative = diff(equation, x)

# Print the original equation and its derivative
print("Original equation:", equation)
print("Derivative:", derivative)
```

Output:
```
Original equation: x**3 + 2*x**2 - 3*x + 5
Derivative: 3*x**2 + 4*x - 3
```

Explanation:

1. We import the necessary functions `symbols` and `diff` from the SymPy library.
2. We define the symbolic variable `x` using `symbols('x')`.
3. We define the equation as an expression involving `x`.
4. We use the `diff` function to differentiate the equation with respect to `x`. The first argument to `diff` is the expression to differentiate, and the second argument is the variable with respect to which the differentiation should be performed.
5. Finally, we print the original equation and its derivative.

In the output, you can see that the original equation is `x^3 + 2*x^2 - 3*x + 5`, and its derivative with respect to `x` is `3*x^2 + 4*x - 3`.

You can modify the equation and the variable(s) as per your requirements. SymPy supports differentiation of complex expressions involving various functions and operations.

Note: Before running the code, you need to install the SymPy library if you haven't already. You can install it using pip:

```
pip install sympy
```To differentiate an equation in Python, we need to use symbolic computation libraries like SymPy. SymPy is a Python library for symbolic mathematics, and it can be used to perform various mathematical operations, including differentiation.

Here's an example code that demonstrates how to differentiate an equation using SymPy:

```python
from sympy import symbols, diff

# Define the symbolic variables
x = symbols('x')

# Define the equation
equation = x**3 + 2*x**2 - 3*x + 5

# Differentiate the equation
derivative = diff(equation, x)

# Print the original equation and its derivative
print("Original equation:", equation)
print("Derivative:", derivative)
```

Output:
```
Original equation: x**3 + 2*x**2 - 3*x + 5
Derivative: 3*x**2 + 4*x - 3
```

Explanation:

1. We import the necessary functions `symbols` and `diff` from the SymPy library.
2. We define the symbolic variable `x` using `symbols('x')`.
3. We define the equation as an expression involving `x`.
4. We use the `diff` function to differentiate the equation with respect to `x`. The first argument to `diff` is the expression to differentiate, and the second argument is the variable with respect to which the differentiation should be performed.
5. Finally, we print the original equation and its derivative.

In the output, you can see that the original equation is `x^3 + 2*x^2 - 3*x + 5`, and its derivative with respect to `x` is `3*x^2 + 4*x - 3`.

You can modify the equation and the variable(s) as per your requirements. SymPy supports differentiation of complex expressions involving various functions and operations.

Note: Before running the code, you need to install the SymPy library if you haven't already. You can install it using pip:

```
pip install sympy
```
"""